Class 45 Pickup Widget
Class 45 Pickup Widget

So in this lecture, we added a pickup widget to our weapon class and we created an overlap function

and bounded to on component begin overlap for our area sphere.

But we only did so on the server.

So the server has control over whether or not the pickup widget can be seen and the lectures to come.