Class 46 Variable Replication
Class 46 Variable Replication

So in this lecture we learned about variable replication and we're replicating a pointer to an a weapon

called overlapping weapon.

And we saw that replication only works one way from server to client.

And we learned about rep notifies, which are functions that get called when a variable gets replicated.

And we saw that the rep notify only gets called on clients as variables never get replicated to the

server only from the server.

And we learned how we can use a condition to allow us to only replicate to the owner of a character.

So rather than replicating a variables value to all clients, we only replicate to the client that's

currently controlling the palm.

In our case, Blaster character.

Great job.

We'll continue in the next lecture.