Class 47. Equipping Weapons
Class 47. Equipping Weapons

So in summary, we created a combat component to handle all functionality related to combat for our

character.

And we made the Blaster character, class a friend, class of the combat component class.

That way, the Blaster character can access all protected and private members on the combat component.

We also added and equip weapon function to the combat component and we're only calling this on the server.

So the server is in charge of equipping weapons.

Now, currently our clients don't have a way to equip weapons.

And so that's going to be one of the next things we figure out how to do.