Class 48 Remote Procedure Calls
So in this lecture we learned about remote procedure calls and how these are functions designed to be

called on one machine and executed on another.

We created a server, our PC designed to be called from a client and executed on the server.

We did this so that we could press the E key on a client and then on the server we could execute the

implementation of that PC to equip the weapon.

Now we saw that some information automatically propagated to the clients, such as attaching the weapon

when we attached the weapon to the character mesh on the server.

That attachment happened on clients.

So we didn't need to do anything about that.

We also saw that setting the owner was a replicated action as the owner variable inherited from the

actor class is replicated and that there is a rep notify for that so we can keep that in mind if we

ever want to override that rep notify.

Now we also replicated our weapon state inNew.

Now that this variables replicated, we can do things in the rep notify for that.

So we went ahead and hid the pickup widget in this rep notify.

So now it'll be hidden on all clients.

When someone picks up a weapon.

And finally, we set collision enabled for the area's fear to no collision so that no more overlap events

will be generated once the weapon is in the equipped state.

And we only needed to do this on the server because overlap events are only generated for our weapon

on the server.

So we're getting closer.

We still need to implement more functionality and we'll continue in the videos to come.