Class - 49 Equipped Animation Pose
So in this lecture, we took care of the equipped animation pose.

We just have the idol pose, but we'll take care of fleshing out our animation blueprint in the videos

to come.

But we now have the correct pose on all machines because we set the equipped weapon variable on the

combat component to be replicated.

And because that variable is replicated, each individual local animation blueprint has that information.

Awesome job.