Class - 50 Crouching
Class - 50 Crouching

So in this lecture we covered the crouch function and how the character has a crouch and an crouch function

built in that takes multiplayer into account.

We saw that our B is crouch variable is replicated and when we call Crouch even from a client and set

on all machines.

So this built in functionality is very convenient for us.

And on top of that, it also takes care of the capsule size and our movement speed, both of which are

variables we can adjust on our character movement component.

So this is very, very convenient.

So we have crouching we haven't yet completed our animation blueprint, but we'll get there.

So this is a great next step.

Awesome job.