Class - 51 Aiming
So in this lecture we handled aiming in multiplayer.

Now we saw that this involved a little bit more work than just calling the crouch function, as crouching

is already built in to the engine and handles multiplayer for us.

So we had to figure out how we could aim in multiplayer and make this work on clients and the server.

So this involved creating a boolean on the combat component called being aiming.

And we also needed to add an aiming boolean to our anim instance so we could change the aiming pose

and we needed to make sure to replicate that variable.

So we set it to replicate it and we saw that this made aiming work for the server as that boolean replicated

to all clients.

But when we aimed on a client we couldn't see that aiming pose on the server or the other client, so

we needed to create an RBC.

The server PC resulted in us changing the boolean value on the server when we aim from a client and

then that Boolean value replicated back to all clients.

We saw that we could also call a server PC on the server and that just results in the server calling

that implementation function.

And that when we call a server RPC on a client that owns the actor that's calling the RTC, it will

be executed on the server.

And we saw some very useful unreal engine documentation on our PCs and I recommend you go back to that

frequently while you're developing multiplayer games.

Awesome job we now have amIng.