Class - 52 Running Blendspace
So in this lecture, we created a running blend space and we duplicated are running animations and rotated

the root bone a bit so that we can have some leaning.

We added a key to those animations and then we bake them into new animations so that those new animations

don't have keys in them.

That way we have that information baked in and we added those to our blend space.

Now our blend space has axis values, and these axis values need to be passed in in our animation blueprint.

So in the next lecture, we're going to go about setting those values in our anim instance so that we

get strafing and cleaning while we're running.

Great job.