Class - 53.Leaning and Strafing
So in this lecture, we took care of cleaning and strafing, and we created two new variables to drive

our running blend space and the animation blueprint.

Now, we calculated these values in the anim instance, and we saw that we could set this up without

making any new replicated variables, which is great because we're not using any more bandwidth to implement

these features.

Now we did see that we needed to take care of orienting rotation to movement and using the controller.

Yeah.

As soon as we equipped a weapon.

And rather than replicating new variables, we took advantage of the fact that we're replicating our

equipped weapon variable.

We simply needed to create a rep, notify for that and disable orienting rotation to movement in the

rep notify as well as the character using the controller.

Yeah.

So this is great.

We're one step closer to having a functional character.