Class - 54 Idle and Jumps
So in this lecture we now have idling and we have jumping and everything is working in multiplayer.

So this is looking great.

We also need to take care of movement while crouching and you may have noticed that we're not able to

aim anymore while we're standing, so we need to add that functionality back in and we'll get there.

We're almost done with the basics of movement for this multiplayer character.

Awesome job.