Class - 55 Crouch Walking
So in this lecture we set up crouching and we now have crouch walking and that's taken care of with

its own blend space.

And we can transition between idle and crouch walking and everything looks good in multiplayer, so

this is great.

But now we don't have any ability to aim and we're going to want to take care of that.

So that'll be in the lectures to come.

Great job.