Class - 56 Aim Walking
So that's going to conclude the video we now have aiming, crouching, and we can aim and crouch.

We can walk and crouch.

We're changing the movement speed when we're aim walking and everything's looking great and it's working

on both clients and the server.

So we now have aim walking in our animation blueprint so we can take care of crouching, standing and

aiming during both of these states, including idle.

So this is great for one step closer to having our basic movement mechanics covered.