Class 57, 58 Creating Aim Offsets and then Applying Aim Offsets
In 57 lecture we created AIm Offsets

So in 58 lecture, we now have our AIM offsets, and we're able to apply them to the character because

we have those values for ya and pitch to hook up to those AIM offsets.

But we did see a strange effect when we play tested in multiplayer.

We're going to discuss that in the next video and and figure out how we can fix it.

Awesome job.