Class - 61 FABRIK IK Some issue here in Animation
So in this lecture we implemented fabric IK for hand placement.

We're using a fabric AK node and specifying the transform for our left hand.

We did this by creating an F transform on the blaster animate instance and adding a socket to our weapon.

And in the anim instance we're getting the transform of that socket and transforming it to bone space

relative to our right hand bone.

Then in the blaster and in BP, we're using that transform and passing it into a fabric K node which

alters our left hand location based on this transform.

And now we can add this socket to any weapon and our fabric AK will place the left hand at the correct

location on the weapon.

This is very versatile and will come in handy as we add more and more different types of weapons.

Awesome job.